Fixed: removing friends only in my account, validation on friendship added

+++ app/controllers/rooms_users_controller.rb
@@ -17,15 +17,18 @@ class RoomsUsersController < ApplicationController
   def destroy
     room_user = RoomsUser.where("user_id = ? AND room_id = ?", params[:user_id], params[:room_id]).first
     room_users_count = RoomsUser.where("room_id = ?", room_user.room_id).count
-    room_user.destroy
-    if (room_users_count -= 1).zero?
-      room = Room.find(params[:room_id])
-      room.destroy
-      #change redirect to main page
-      redirect_to root_path
+    if(current_user.id == params[:user_id].to_i)
+      room_user.destroy
+      if (room_users_count -= 1).zero?
+        room = Room.find(params[:room_id])
+        room.destroy
+        #change redirect to main page
+        redirect_to root_path
+      end
     else
-      redirect_to room_path(params[:room_id])
+      render json: {:drop_user_id => params[:user_id], :cur_user_id => current_user.id}
     end
+    #redirect_to room_path(params[:room_id])
   end
 
 end
